refactor: move is_target_asn function to utils and update imports
