+++ src/autodialer/utils/__init__.py
@@ -1,3 +1,3 @@
-from .validate_asn import normalize_asn, validate_asn
+from .validate_asn import is_target_asn, normalize_asn, validate_asn
 
-__all__ = ["validate_asn", "normalize_asn"]
+__all__ = ["validate_asn", "normalize_asn", "is_target_asn"]
